Refactor digitalFormPatterns specs
