+++ src/applications/form-renderer/tests/unit/utils/formConfig.unit.spec.js
@@ -6,8 +6,7 @@ import { expect } from 'chai';
 import sinon from 'sinon';
 import * as IntroductionPage from 'applications/form-renderer/containers/IntroductionPage';
 import { render } from '@testing-library/react';
-import * as webComponentPatterns from 'platform/forms-system/src/js/web-component-patterns';
-import * as addressPatterns from 'platform/forms-system/src/js/web-component-patterns/addressPattern';
+import * as digitalFormPatterns from 'applications/form-renderer/utils/digitalFormPatterns';
 import { normalizedForm } from '../../../_config/formConfig';
 import { createFormConfig, selectSchemas } from '../../../utils/formConfig';
 import manifest from '../../../manifest.json';
@@ -81,191 +80,60 @@ describe('createFormConfig', () => {
 });
 
 describe('selectSchemas', () => {
-  const findChapterByType = type =>
-    normalizedForm.chapters.find(chapter => chapter.type === type);
+  const [
+    nameAndDob,
+    ,
+    identificationInfo,
+    address,
+    phoneAndEmail,
+  ] = normalizedForm.chapters;
+  let spy;
 
   context('with Name and Date of Birth pattern', () => {
-    let dobIncluded;
-
     beforeEach(() => {
-      dobIncluded = selectSchemas(
-        findChapterByType('digital_form_name_and_date_of_bi'),
-      );
+      spy = sinon.spy(digitalFormPatterns, 'digitalFormNameAndDoB');
     });
 
-    it('contains fullName', () => {
-      expect(dobIncluded.schema.properties.fullName).to.eq(
-        webComponentPatterns.fullNameSchema,
-      );
-      expect(dobIncluded.uiSchema.fullName).to.not.eq(undefined);
-    });
+    it('calls digitalFormNameAndDob', () => {
+      selectSchemas(nameAndDob);
 
-    context('when includeDateOfBirth is true', () => {
-      it('contains dateOfBirth', () => {
-        expect(dobIncluded.schema.properties.dateOfBirth).to.eq(
-          webComponentPatterns.dateOfBirthSchema,
-        );
-        expect(dobIncluded.uiSchema.dateOfBirth).to.not.eq(undefined);
-      });
-    });
-
-    context('when includeDateOfBirth is false', () => {
-      it('does not contain dateOfBirth', () => {
-        const nameOnlyChapter = {
-          id: 158254,
-          chapterTitle: 'Name only chapter',
-          type: 'digital_form_name_and_date_of_bi',
-          pageTitle: 'Name and Date of Birth',
-          additionalFields: {
-            includeDateOfBirth: false,
-          },
-        };
-        const nameOnly = selectSchemas(nameOnlyChapter);
-
-        expect(nameOnly.schema.properties.dateOfBirth).to.eq(undefined);
-        expect(nameOnly.uiSchema.dateOfBirth).to.eq(undefined);
-      });
+      expect(spy.calledWith(nameAndDob)).to.eq(true);
     });
   });
 
   context('with Identification Information pattern', () => {
-    let vetIdOnly;
-
     beforeEach(() => {
-      vetIdOnly = selectSchemas(
-        findChapterByType('digital_form_identification_info'),
-      );
+      spy = sinon.spy(digitalFormPatterns, 'digitalFormIdentificationInfo');
     });
 
-    it('contains veteranId', () => {
-      expect(vetIdOnly.schema.properties.veteranId).to.eq(
-        webComponentPatterns.ssnOrVaFileNumberSchema,
-      );
-      expect(vetIdOnly.uiSchema.veteranId).to.not.eq(undefined);
-    });
+    it('calls digitalFormIdentificationInfo', () => {
+      selectSchemas(identificationInfo);
 
-    context('when includeServiceNumber is true', () => {
-      it('includes serviceNumber', () => {
-        const serviceNumberChapter = {
-          id: 158255,
-          chapterTitle: 'Service Number Included',
-          type: 'digital_form_identification_info',
-          pageTitle: 'Identification Information',
-          additionalFields: {
-            includeServiceNumber: true,
-          },
-        };
-        const serviceNumberIncluded = selectSchemas(serviceNumberChapter);
-
-        expect(serviceNumberIncluded.schema.properties.serviceNumber).to.eq(
-          webComponentPatterns.serviceNumberSchema,
-        );
-        expect(serviceNumberIncluded.uiSchema.serviceNumber).to.not.eq(
-          undefined,
-        );
-      });
-    });
-
-    context('when includeServiceNumber is false', () => {
-      it('does not include serviceNumber', () => {
-        expect(vetIdOnly.schema.properties.serviceNumber).to.eq(undefined);
-        expect(vetIdOnly.uiSchema.serviceNumber).to.eq(undefined);
-      });
+      expect(spy.calledWith(identificationInfo)).to.eq(true);
     });
   });
 
   context('with Address pattern', () => {
-    let addressSpy;
-    let noMiliarySpy;
-
     beforeEach(() => {
-      addressSpy = sinon.spy(addressPatterns, 'addressSchema');
-      noMiliarySpy = sinon.spy(addressPatterns, 'addressNoMilitarySchema');
+      spy = sinon.spy(digitalFormPatterns, 'digitalFormAddress');
     });
 
-    afterEach(() => {
-      addressSpy.restore();
-      noMiliarySpy.restore();
-    });
+    it('calls digitalFormAddress', () => {
+      selectSchemas(address);
 
-    context('when militaryAddressCheckbox is true', () => {
-      it('calls addressSchema', () => {
-        const includeMilitary = {
-          id: 158256,
-          chapterTitle: 'Includes military addresses',
-          type: 'digital_form_address',
-          pageTitle: 'Address',
-          additionalFields: {
-            militaryAddressCheckbox: true,
-          },
-        };
-        const schemas = selectSchemas(includeMilitary);
-
-        expect(addressSpy.calledOnce).to.eq(true);
-        expect(schemas.uiSchema.address).to.not.eq(undefined);
-      });
-    });
-
-    context('when militaryAddressCheckbo is false', () => {
-      it('calls addressNoMilitarySchema', () => {
-        const schemas = selectSchemas(
-          findChapterByType('digital_form_address'),
-        );
-
-        expect(noMiliarySpy.calledOnce).to.eq(true);
-        expect(schemas.uiSchema.address).to.not.eq(undefined);
-      });
+      expect(spy.calledWith(address)).to.eq(true);
     });
   });
 
   context('with Phone and Email Address pattern', () => {
-    const includeEmail = {
-      id: 158256,
-      chapterTitle: 'Email address included',
-      type: 'digital_form_phone_and_email',
-      pageTitle: 'Phone and email address',
-      additionalFields: {
-        includeEmail: true,
-      },
-    };
-    const phoneOnly = findChapterByType('digital_form_phone_and_email');
-
-    it('includes homePhone', () => {
-      const schemas = selectSchemas(includeEmail);
-
-      expect(schemas.schema.properties.homePhone).to.eq(
-        webComponentPatterns.phoneSchema,
-      );
-      expect(schemas.uiSchema.homePhone).to.not.eq(undefined);
+    beforeEach(() => {
+      spy = sinon.spy(digitalFormPatterns, 'digitalFormPhoneAndEmail');
     });
 
-    it('includes mobilePhone', () => {
-      const schemas = selectSchemas(phoneOnly);
+    it('calls digitalFormPhoneAndEmail', () => {
+      selectSchemas(phoneAndEmail);
 
-      expect(schemas.schema.properties.mobilePhone).to.eq(
-        webComponentPatterns.phoneSchema,
-      );
-      expect(schemas.uiSchema.mobilePhone).to.not.eq(undefined);
-    });
-
-    context('when includeEmail is true', () => {
-      it('includes emailAddress', () => {
-        const schemas = selectSchemas(includeEmail);
-
-        expect(schemas.schema.properties.emailAddress).to.eq(
-          webComponentPatterns.emailSchema,
-        );
-        expect(schemas.uiSchema.emailAddress).to.not.eq(undefined);
-      });
-    });
-
-    context('when includeEmail is false', () => {
-      it('does not include emailAddress', () => {
-        const schemas = selectSchemas(phoneOnly);
-
-        expect(schemas.schema.properties.emailAddress).to.eq(undefined);
-        expect(schemas.uiSchema.emailAddress).to.eq(undefined);
-      });
+      expect(spy.calledWith(phoneAndEmail)).to.eq(true);
     });
   });
 });
